refactor: config type relate to type of library
