+++ src/components/Avatar/AvatarPersonal.tsx
@@ -1,7 +1,15 @@
-import { Avatar } from '@mui/material';
+import { Avatar, AvatarProps } from '@mui/material';
 
-const AvatarPersonal = ({ srcImage }: { srcImage: string }): JSX.Element => {
-  return <Avatar src={srcImage} alt='' sx={{ width: 300, height: 300 }} />;
+interface AvatarPersonalProps extends AvatarProps {}
+
+const AvatarPersonal = ({ src }: AvatarPersonalProps) => {
+  return (
+    <Avatar
+      src={src}
+      alt='This is a personal avatar'
+      sx={{ width: 300, height: 300 }}
+    />
+  );
 };
 
 export default AvatarPersonal;
